feat: star canvas easter egg refactored
